move two_pp_cluster to schedules
